Add support for aggregate distinct for SUM and AVG.
We found that DISTINCT is not supported for the datasource aggregate pushdown patch.
Examples of the use of DISTINCT:   SUM(DISTINCT column) or AVG(DISTINCT column).

This adds DISTINCT support for both SUM and AVG.  Previously spark did not give
the datasource any hint that the SUM/AVG was a distinct, and as a result the datasource
treated both cases as a regular SUM/AVG.  With this PR, spark now passes information
on distinct to the datasource, so it can form the appropriate query.

We also updated the JDBCV2Suite tests with SUM and AVG distinct cases.